backend/requirements.txt
# ========================================
# ALGORIST BACKEND DEPENDENCIES
# ========================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
python-dotenv==1.0.1

# Data Fetching & Processing
yahooquery==2.3.7
requests==2.32.3
pandas==2.3.3
beautifulsoup4==4.14.3
lxml==6.0.2

# Turkish Market Data
tefas-crawler==0.2.1

# Scheduling & Background Jobs
APScheduler==3.10.4

# Type Checking Support (Professional Development)
pandas-stubs==2.3.3.260113
types-requests==2.32.4.20260107
types-python-dateutil==2.9.0.20251115
types-pytz==2025.2.0.20251108

# Timezone & Date Support
pytz==2025.2
tzdata==2025.3
tzlocal==5.3.1
python-dateutil==2.9.0.post0
